Fix package name.
