+++ datastore/datastore-utils/src/test/resources/org/orbisgis/datastore/datastoreutils/landcover2000.shx
