+++ datastore/h2gis/src/main/groovy/org/orbisgis/datastore/h2gis/H2GISDataStoreParamsCst.groovy
@@ -39,7 +39,7 @@ package org.orbisgis.datastore.h2gis
 import groovy.transform.BaseScript
 import groovy.transform.Field
 import org.h2gis.geotools.H2GISDataStoreFactory
-import org.orbisgis.datastore.utils.JDBCDataStoreParamsCst
+import org.orbisgis.datastore.jdbcutils.JDBCDataStoreParamsCst
 
 /**
  * Groovy script used as Closure DelegateTo on opening a H2GIS JDBCDataStore.
