+++ datastore/jdbc-utils/src/main/resources/META-INF/groovy/org.codehaus.groovy.runtime.ExtensionModule
@@ -1,4 +1,4 @@
 moduleName=OrbisGIS Groovy datastore extension
 moduleVersion=1.0
-extensionClasses=org.orbisgis.datastore.utils.JDBCDataStoreUtils
+extensionClasses=org.orbisgis.datastore.jdbcutils.JDBCDataStoreUtils
 staticExtensionClasses=
